utilizing DbCommand and removing lots of superfluous code from children
